5062 - Move all 2021-2022 files into their own package
