+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/Blue20SecAuto.java
@@ -27,11 +27,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -40,38 +39,34 @@ import com.qualcomm.robotcore.util.ElapsedTime;
  * This file illustrates the concept of driving a path based on time.
  * It uses the common Pushbot hardware class to define the drive on the robot.
  * The code is structured as a LinearOpMode
- * <p>
+ *
  * The code assumes that you do NOT have encoders on the wheels,
- * otherwise you would use: PushbotAutoDriveByEncoder;
- * <p>
- * The desired path in this example is:
- * - Drive forward for 3 seconds
- * - Spin right for 1.3 seconds
- * - Drive Backwards for 1 Second
- * - Stop and close the claw.
- * <p>
- * The code is written in a simple form with no optimizations.
- * However, there are several ways that this type of sequence could be streamlined,
- * <p>
+ *   otherwise you would use: PushbotAutoDriveByEncoder;
+ *
+ *   The desired path in this example is:
+ *   - Drive forward for 3 seconds
+ *   - Spin right for 1.3 seconds
+ *   - Drive Backwards for 1 Second
+ *   - Stop and close the claw.
+ *
+ *  The code is written in a simple form with no optimizations.
+ *  However, there are several ways that this type of sequence could be streamlined,
+ *
  * Use Android Studios to Copy this Class, and Paste it into your team's code folder with a new name.
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@Autonomous(name = "Blue Freight", group = "Autonomous")
+@Autonomous(name="Blue20SecAuto", group="Autonomous")
 //@Disabled
-public class BlueFreight extends LinearOpMode {
-
+public class Blue20SecAuto extends LinearOpMode {
     /* Declare OpMode members. */
     private ElapsedTime runtime = new ElapsedTime();
     private DcMotor leftTop;
     private DcMotor leftBottom;
     private DcMotor rightTop;
     private DcMotor rightBottom;
-    //private DcMotor duckWheel;
     private DcMotor arm;
     private Servo claw;
-    //private CRServo duckWheel;
-
 
     public void setup() {
 
@@ -79,10 +74,8 @@ public class BlueFreight extends LinearOpMode {
         leftBottom = hardwareMap.get(DcMotor.class, "leftBottom");
         rightTop = hardwareMap.get(DcMotor.class, "rightTop");
         rightBottom = hardwareMap.get(DcMotor.class, "rightBottom");
-        //duckWheel = hardwareMap.get(DcMotor.class, "duckWheel");
         arm = hardwareMap.get(DcMotor.class, "arm");
         claw = hardwareMap.get(Servo.class, "claw");
-        //duckWheel = hardwareMap.get(CRServo.class, "duckWheel");
 
         leftTop.setDirection(DcMotor.Direction.REVERSE);
         leftBottom.setDirection(DcMotor.Direction.REVERSE);
@@ -94,8 +87,6 @@ public class BlueFreight extends LinearOpMode {
         leftBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightTop.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-
-//        arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         // Send telemetry message to signify robot waiting;
         telemetry.addData("Status", "Initiated");    //
@@ -111,37 +102,43 @@ public class BlueFreight extends LinearOpMode {
         // Wait for the game to start (driver presses PLAY)
         waitForStart();
 
-        // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
-
-        // Step 1:  Drive forward for 3 seconds
-
         claw.setPosition(1.0);
-        sleep(1000);
+        sleep(20000);
 
-        //ducks(1);
+        // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
 
         raiseArm(150);
 
-//        leftTop.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-//        leftBottom.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-//        rightTop.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-//        rightBottom.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        strafeToPosition(890);
 
-        int drivePosition = 860;
+        raiseArm(-145);
 
-        driveToPosition(drivePosition);
+        // Step 1:  Drive forward for 3 seconds
+        /*arm.setPower(0.3);
+        runtime.reset();
+        while (opModeIsActive() && (runtime.seconds() < 1.0)) {
+            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
+            telemetry.update();
+        }
 
-        sleep(500);
+        strafeRight(0.3);
 
-        claw.setPosition(0);
-        sleep(500);
+        runtime.reset();
+        while (opModeIsActive() && (runtime.seconds() < 4.0)) {
+            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
+            telemetry.update();
+        }
 
-        raiseArm(50);
-        driveToPosition(-125);
+        // Step 2:  Stop
+        strafeRight(0); //Stops motors
+        arm.setPower(0);
 
-        strafeToPosition(1650);
 
+        telemetry.addData("Path", "Complete");
+        telemetry.update();
+        sleep(1000);*/
     }
+
 
     private void driveToPosition(int drivePosition) {
         int distance = leftBottom.getCurrentPosition() + drivePosition;
@@ -163,16 +160,15 @@ public class BlueFreight extends LinearOpMode {
 //            telemetry.update();
 
         }
-
         // 6. Turn off the motor
         drive(0, 0);
     }
 
     private void strafeToPosition(int strafePosition) {
-        leftTop.setTargetPosition(leftTop.getCurrentPosition() + strafePosition);
-        leftBottom.setTargetPosition(leftBottom.getCurrentPosition() - strafePosition);
-        rightTop.setTargetPosition(rightTop.getCurrentPosition() - strafePosition);
-        rightBottom.setTargetPosition(rightBottom.getCurrentPosition() + strafePosition);
+        leftTop.setTargetPosition(leftTop.getCurrentPosition() - strafePosition);
+        leftBottom.setTargetPosition(leftBottom.getCurrentPosition() + strafePosition);
+        rightTop.setTargetPosition(rightTop.getCurrentPosition() + strafePosition);
+        rightBottom.setTargetPosition(rightBottom.getCurrentPosition() - strafePosition);
 
         leftBottom.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         leftTop.setMode(DcMotor.RunMode.RUN_TO_POSITION);
@@ -213,7 +209,7 @@ public class BlueFreight extends LinearOpMode {
 
     private void raiseArm(int position) {
 //        arm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
- //       arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        //       arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         telemetry.addData("Distance", position);    //
         telemetry.addData("Starting Arm position", arm.getCurrentPosition());    //
@@ -236,63 +232,5 @@ public class BlueFreight extends LinearOpMode {
         telemetry.addData("Fallen Arm position", arm.getCurrentPosition());    //
         telemetry.update();
         sleep(1000);
-    }
-
-    /*private void ducks(int duckSpeed) {
-        duckWheel.setPower(duckSpeed);
-    }*/
-
-    /*leftTop.setPower(0.4);
-    leftBottom.setPower(0.4);
-    rightTop.setPower(0.4);
-    rightBottom.setPower(0.4);
-    while (opModeIsActive() && (runtime.seconds() < 1.25)) {
-        telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-        telemetry.update();
-    }
-
-    strafeRight(0.3);
-
-    runtime.reset();
-    while (opModeIsActive() && (runtime.seconds() < 1.0)) {
-        telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-        telemetry.update();
-    }
-
-    // Step 2:  Stop
-    strafeRight(0); //Stops motors
-
-
-    telemetry.addData("Path", "Complete");
-    telemetry.update();
-    sleep(1000);
-}
-*/
-    public void moveLeft(double speed) {
-        telemetry.addData("moveLeft speed", speed);
-
-        double rb_speed_var = 1;
-        if (speed < 0) {
-            rb_speed_var = -1.0 / speed;
-        }
-        double rb_speed = speed * rb_speed_var;
-        telemetry.addData("Strafe speed", speed);
-        telemetry.addData("rb_speed_var", rb_speed_var);
-        telemetry.addData("rb_speed", rb_speed);
-
-        leftTop.setPower(-speed);
-        leftBottom.setPower(speed);
-        rightTop.setPower(speed);
-        rightBottom.setPower(-rb_speed);
-
-        telemetry.addData("LT Speed", leftTop.getPower());
-        telemetry.addData("LB Speed", leftBottom.getPower());
-        telemetry.addData("RT Speed", rightTop.getPower());
-        telemetry.addData("RB Speed", rightBottom.getPower());
-    }
-
-    public void moveRight(double speed) {
-        telemetry.addData("moveRight speed", speed);
-        moveLeft(-speed);
     }
 }
