+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/trish_sandbox/DriveTrish.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.trish_sandbox;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022.trish_sandbox;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
