+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/trish_sandbox/RobotConfigTrish.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.trish_sandbox;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022.trish_sandbox;
 
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
